# Conflicts:
#   util-ktx/RELEASE_NOTES.md

+++ app-migration/sample/src/main/java/ru/surfstudio/android/app/migration/sample/app/initialization/migration/AppMigrationStorageImpl.kt
@@ -10,13 +10,13 @@ class AppMigrationStorageImpl : AppMigrationStorage {
 
     companion object {
 
-        private const val MIGRATION_FOR_EACH = -1
+        private const val MIGRATION_FOR_EACH = -1L
 
         private val MIGRATIONS = arrayOf(
                 //вызывается для каждого обновления
                 object : AppMigration(MIGRATION_FOR_EACH) {
 
-                    override fun execute(oldVer: Int, newVer: Int) {
+                    override fun execute(oldVer: Long, newVer: Long) {
                         Logger.d("Executing app migration for each update")
                         apply() //нет проверки версий
                     }
@@ -26,7 +26,7 @@ class AppMigrationStorageImpl : AppMigrationStorage {
                     }
                 },
 
-                object : AppMigration(2) {
+                object : AppMigration(2L) {
                     override fun apply() {
                         //здесь должен быть код миграции на версию 2
                     }
